backend/requirement.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Werkzeug==3.0.1
openpyxl==3.1.2
boto3==1.34.69
requests==2.32.3
